fee collection modal changes
